Fix P3-11 (trig precision) and P3-12 (lua_gc variadic cast)
P3-11: Replace (long long) cast in _reduce_angle() with fmod().
Also fix fmod() itself in heaven_stubs.c — the old implementation
had the same precision loss for |x/y| > 2^53. New implementation
uses iterative power-of-2 subtraction for huge quotients.

P3-12: Add explicit `as c_int` casts on lua_gc() variadic args
for portability to architectures where c_int != i32.

All 12 audit findings are now resolved.

+++ kernel/src/lua/mod.rs
@@ -17,7 +17,7 @@ pub mod repl;
 use ::alloc::string::String;
 use ::alloc::vec::Vec;
 
-use core::ffi::c_void;
+use core::ffi::{c_int, c_void};
 
 use ffi::*;
 
@@ -55,7 +55,7 @@ pub fn run_string(code: &str, name: &str) -> Result<(), String> {
         luaL_openlibs(L);
 
         // 3. Configure GC for incremental mode with small steps
-        lua_gc(L, LUA_GCINC, 100, 200, 10);
+        lua_gc(L, LUA_GCINC, 100 as c_int, 200 as c_int, 10 as c_int);
 
         // 4. Register OSqlite builtins
         builtins::register_builtins(L);
